Fix creation of result file in Python 3 (#42)
* Fix creation of result file in Python 3

The code relied on evals inside list comprehensions to be executed in
the surrounding scope. This is not the case in Python 3.

The function `unicode` doesn't exist in Python 3 either, it is now
replaced with `str` on Python 3.

* Add parameter to control header printing

As discussed in https://github.com/sequitur-g2p/sequitur-g2p/pull/42#issuecomment-461860406
this change keeps breaking changes to a minimum.