Setup publishing wheels through GitHub actions (#69)
* Setup publishing wheels through GitHub actions

I wasn't able to get TravisCI to work very well for this, but it wasn't
difficult to switch to GitHub actions. This seems to build wheels for
Windows, macOS, and Ubuntu just fine.

To get the wheels in PyPI (see #67), I think we will just need to set
the environment variables `PYPI_API_TOKEN` and `TEST_PYPI_API_TOKEN` and
uncomment the last few steps. Then new releases and tags should be
automatically built and sent to PyPI.

We could also switch (or add) a trigger for releases if preferred:

```
on:
  release:
    types: [created]
```

* Empty commit to trigger build